feat(search): added typings and enhanced query resolution
