+++ utils/i18n.ts
@@ -3,6 +3,7 @@ import CONTENT_EN from "../i18n/content.en.json"
 import CONTENT_JA from "../i18n/content.ja.json"
 import UI_EN from "../i18n/ui.en.json"
 import UI_JA from "../i18n/ui.ja.json"
+import type { Content } from "search"
 
 export type Locale = (typeof locales)[number]
 
@@ -18,4 +19,4 @@ export const ui = { ja: UI_JA, en: UI_EN }
 export const contents = { ja: CONTENT_JA, en: CONTENT_EN }
 
 export const getUI = (locale: Locale) => ui[locale]
-export const getContents = (locale: Locale) => contents[locale]
+export const getContents = (locale: Locale) => contents[locale] as Content[]
